Add files via upload
Hi. The variables v1, v2, v3 all the way to v50 should be named. The ones that are going to repeat are lat, long, speed, long acc, lat accel, and time. The ones that are constant are ID, type, distance, speed. I was doing an if-statement to get the variables to store automatically. But I realized this approach doesn't work, because then the repeating variables would constantly be assigned to a new V variable. A team member used the possibility of using an array?